+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+* [L10n] `to_xml`:
+  * Prints info about written diffs.
+  * Do not create empty diff file.
+* [L10n] `import_xml`:
+  * Prints info about imported diffs.
+  * Handle diff file with locale at the end.
+  * Add documentation about import diffs.
 
-* [L10n] Prints info about written diffs.
-* [L10n] Do not create empty diff file.
 ## 1.6.1
 
 * [Release] Support for provide scrips to run before release.